google_appscript: build JS(.gs) code to periodically fetch & update spreadsheet
Implements code to periodically fetch data from the API and then update
the spreadsheet with dependency versions and it will be possible to have
the scripts running periodically e.g. every 1 hour.
The API server isn't yet deployed though.

This change also changed the response returned from fetching from /
to be a map indexed by chain_name instead of it being a slice of
*ChainSchema objects; this allows quick look ups in the JS script
which keeps the code simple & compact.

Fixes #4