chainparse: always use latest go.mod since chain-registry not maintained
This code uses the always live Github version of each repository
because sadly the content on the chain-registry is heavily unmaintained.
This change performs a shallow git clone to only get the default branch
from `.git/HEAD` => `/refs/<DEFAULT_BRANCH>` to then go fetch the live
code. It is so much better to get the live data that try to guess what
was deployed and in the chain registry since we've got deadlinks like:
* https://github.com/imversed/imversed

Fixes #9